refactor(packages): Move CSL support module unter bibtex package
