+++ packages/bibtex/csl/locales/locales-fr-FR.xml
