+++ packages/bibtex/csl/styles/chicago-author-date-fr.csl
